MdeModulePkg: ScsiDiskDxe: fix to support EFI_ERASE_BLOCK_PROTOCOL
If bit TPZ and bit TPRZ are set, the erase feature is implemented.
If bit TPZ is set and bit TPRZ is clear, the discard feature is
implemented. And discard is a non-secure variant of the erase
functionality.

So the detecting operation of EFI_ERASE_BLOCK_PROTOCOL, we should
consider to support both functionality. Since discard functionality is
only supported in some UFS devices.

And both of these two features are relied on UNMAP command.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Haojian Zhuang <haojian.zhuang@linaro.org>